backend/requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Redis and Celery
redis==5.0.1
celery==5.3.4

# AI Providers
openai==1.3.7
google-generativeai==0.3.1
anthropic==0.7.7

# Text Extraction
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
chardet==5.2.0

# NLP and ML
nltk==3.8.1
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
structlog==23.2.0
httpx==0.25.2

# PDF Generation
reportlab==4.0.7
weasyprint==60.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
